Split t/git.t into sections aligned with libgit2 test suite
The single git.t file is not well named, and a lot of unrelated tests
are included in the same file.  However internally it is well
structured, so perhaps it makes more sense to structure it along the
lines of the libgit2 library.  This is a simple split of the git.t
file into four separate tests.  The signature tests are included in
04-commit.t, but we'll move them to 04-signature.t for the moment.